Bump Plugin Versions

+++ app/plugins/updraftplus/methods/dropbox.php
@@ -17,12 +17,14 @@ if (!is_array(UpdraftPlus_Options::get_updraft_option('updraft_dropbox'))) {
 		'tk_request_token' => UpdraftPlus_Options::get_updraft_option('updraft_dropboxtk_request_token'),
 		'tk_access_token' => UpdraftPlus_Options::get_updraft_option('updraft_dropboxtk_access_token'),
 	);
-	UpdraftPlus_Options::update_updraft_option('updraft_dropbox', $opts);
-	UpdraftPlus_Options::delete_updraft_option('updraft_dropbox_appkey');
-	UpdraftPlus_Options::delete_updraft_option('updraft_dropbox_secret');
-	UpdraftPlus_Options::delete_updraft_option('updraft_dropbox_folder');
-	UpdraftPlus_Options::delete_updraft_option('updraft_dropboxtk_request_token');
-	UpdraftPlus_Options::delete_updraft_option('updraft_dropboxtk_access_token');
+	if (serialize($opts) != 'a:5:{s:6:"appkey";N;s:6:"secret";N;s:6:"folder";N;s:16:"tk_request_token";N;s:15:"tk_access_token";N;}') {
+		UpdraftPlus_Options::update_updraft_option('updraft_dropbox', $opts);
+		UpdraftPlus_Options::delete_updraft_option('updraft_dropbox_appkey');
+		UpdraftPlus_Options::delete_updraft_option('updraft_dropbox_secret');
+		UpdraftPlus_Options::delete_updraft_option('updraft_dropbox_folder');
+		UpdraftPlus_Options::delete_updraft_option('updraft_dropboxtk_request_token');
+		UpdraftPlus_Options::delete_updraft_option('updraft_dropboxtk_access_token');
+	}
 }
 
 class UpdraftPlus_BackupModule_dropbox {
@@ -407,7 +409,7 @@ class UpdraftPlus_BackupModule_dropbox {
 
 			<?php
 
-				$defmsg = '<tr class="updraftplusmethod dropbox"><td></td><td><strong>'.__('Need to use sub-folders?','updraftplus').'</strong> '.__('Backups are saved in','updraftplus').' apps/UpdraftPlus. '.__('If you back up several sites into the same Dropbox and want to organise with sub-folders, then ','updraftplus').'<a href="http://updraftplus.com/shop/">'.__("there's an add-on for that.",'updraftplus').'</a></td></tr>';
+				$defmsg = '<tr class="updraftplusmethod dropbox"><td></td><td><strong>'.__('Need to use sub-folders?','updraftplus').'</strong> '.__('Backups are saved in','updraftplus').' apps/UpdraftPlus. '.__('If you back up several sites into the same Dropbox and want to organise with sub-folders, then ','updraftplus').'<a href="https://updraftplus.com/shop/">'.__("there's an add-on for that.",'updraftplus').'</a></td></tr>';
 
 				echo apply_filters('updraftplus_dropbox_extra_config', $defmsg); ?>
